+++ src/modules/demux/options_parser_demux.cpp
@@ -51,7 +51,7 @@ bool options_parser_demux::parse( int argc, char ***argv, options *opts )
         ( "index1", po::value<std::string>()
                      ->required()
                      ->notifier( [&]( const std::string &vals ) {
-                             opts_demux->set_info( &options_demux::f_index_data,
+                             opts_demux->set_info( &options_demux::index1_data,
                                                     vals
                                                  );
                                                                 }
@@ -74,7 +74,7 @@ bool options_parser_demux::parse( int argc, char ***argv, options *opts )
         ( "index2", po::value<std::string>()
                      ->default_value( "0,0,0" )
                      ->notifier( [&]( const std::string &vals ) {
-                             opts_demux->set_info( &options_demux::r_index_data,
+                             opts_demux->set_info( &options_demux::index2_data,
                                                     vals
                                                  );
                                                                 }
@@ -107,9 +107,22 @@ bool options_parser_demux::parse( int argc, char ***argv, options *opts )
           "Note: When this mode is used, the name of the aggregate sequence will be the sequence that was a result of the translation. "
           "Therefore, this mode is most appropriate for use with reference-independent demultiplexing.\n"
         )
-        ( "samplelist,s", po::value<std::string>( &opts_demux->samplelist_fname )->required(), "A tab-delimited list of samples, one sample per line. "
-          "The first item in each tab-delimited line is \"--index1\", the second (if included) is the \"--index2\", and the third is the samplename. "
-          "\"--index2\" can be omitted if only using a single index.\n"
+        ( "samplelist,s", po::value<std::string>( &opts_demux->samplelist_fname )->required(), "A tab-delimited list of samples with a header row "
+          "and one sample per line. This file must contain at least one index column and one sample name column. Multiple index columns may be included. "
+          "This file can also include additional columns that will not be used for the demultiplexing. "
+          "Specify which columns to use with the \"--sname\", \"--sindex1\", and \"--sindex2\" flags.\n"
+        )
+        (
+          "sname", po::value<std::string>( &opts_demux->samplename )->default_value( "SampleName" ),
+          "Used to specify the header for the sample name column in the samplelist. By default \'SampleName\' is set as the column header name.\n"
+        )
+        (
+          "sindex1", po::value<std::string>( &opts_demux->sample_idx1 )->default_value( "Index1" ),
+          "Used to specify the header for the index 1 column in the samplelist. By default \'Index1\' is set as the column header name.\n"
+        )
+        (
+          "sindex2", po::value<std::string>( &opts_demux->sample_idx2 )->default_value( "Index2" ),
+          "Used to specify the header for the index 2 column in the samplelist. By default \'Index2\' is set as the column header name.\n"
         )
         ( "diagnostic_info,d", po::value<std::string>( &opts_demux->diagnostic_fname )->default_value( "" ),
           "Include this flag with an output file name to collect diagnostic information on read pair matches in map. The file will be formated with tab delimited "